관리자 Controller 구현
관리자 Controller 구현 및 기존 Controller 수정

+++ src/main/java/capstone3d/Server/api/RoomApiController.java
@@ -16,7 +16,7 @@ public class RoomApiController {
     private final RoomRepository roomRepository;
 
     @GetMapping("/room/{id}")
-    public RoomUrlDto RoomUrl(@PathVariable("id") int id) {
+    public RoomUrlDto RoomUrl(@PathVariable("id") Long id) {
         Room findRoom = roomRepository.findById(id);
         User user = findRoom.getUser();
         String business_name = user.getBusiness_name();
